refactor(shared): move supabase-auth and supabase-server into adapters/supabase/
Two adapter files lived as siblings of the adapters/supabase/ directory
without a clear reason. Moved to adapters/supabase/auth.ts and
adapters/supabase/server.ts so all Supabase adapters live in one place.

External callers go via @fyndstigen/shared so no public-API change.

Closes #90
