+++ packages/shared/src/adapters/index.ts
@@ -1,6 +1,6 @@
 import type { SupabaseClient } from '@supabase/supabase-js'
-import { createSupabaseAuth } from './supabase-auth'
-import { createSupabaseServerData } from './supabase-server'
+import { createSupabaseAuth } from './supabase/auth'
+import { createSupabaseServerData } from './supabase/server'
 import { createSupabaseFleaMarkets, createSupabaseSearch, createSupabaseMarketTables } from './supabase/flea-markets'
 import { createSupabaseBookings } from './supabase/bookings'
 import { createSupabaseRoutes } from './supabase/routes'
@@ -12,8 +12,8 @@ import type { BookingRepository } from '../ports/bookings'
 import type { RouteRepository } from '../ports/routes'
 import type { ProfileRepository, OrganizerRepository } from '../ports/profiles'
 
-export { createSupabaseAuth } from './supabase-auth'
-export { createSupabaseServerData } from './supabase-server'
+export { createSupabaseAuth } from './supabase/auth'
+export { createSupabaseServerData } from './supabase/server'
 export { createSupabaseFleaMarkets, createSupabaseSearch, createSupabaseMarketTables } from './supabase/flea-markets'
 export { createSupabaseBookings } from './supabase/bookings'
 export { createSupabaseRoutes } from './supabase/routes'
